+++ src/matrix_sum.h
@@ -8,9 +8,9 @@
 using namespace Rcpp;
 
 // converts Rcpp matrixes into std vectors
-std::vector<std::vector<float>> matrix_to_vectors(Rcpp::NumericMatrix mat);
+std::vector<std::vector<double>> matrix_to_vectors(Rcpp::NumericMatrix mat);
 
 // converts std vectors into Rcpp matrixes
-Rcpp::NumericMatrix vectors_to_matrix(std::vector<std::vector<float>> vecs);
+Rcpp::NumericMatrix vectors_to_matrix(std::vector<std::vector<double>> vecs);
 
 #endif // MATRIX_SUM_H
